gcp-free-tier-api/requirements.txt
flask==3.1.3
flask-cors==6.0.0
google-cloud-bigquery==3.14.0
pandas==2.1.4
gunicorn==22.0.0
